+++ src/index.ts
@@ -1,45 +1,61 @@
 import joplin from 'api';
-import { MenuItemLocation, Path } from 'api/types';
+import { MenuItemLocation, Path, SettingItemType } from 'api/types';
+import { ChangeEvent } from 'api/JoplinSettings';
+import { LayoutType, layoutDesc } from './helpers';
 
 joplin.plugins.register({
   onStart: async function () {
     const COMMANDS = joplin.commands;
-    const DATA = joplin.data;
     const SETTINGS = joplin.settings;
     const WORKSPACE = joplin.workspace;
 
+    //#region SETTINGS
+
+    await SETTINGS.registerSection('persistent.layout.settings', {
+      label: 'Persistent Layout',
+      iconName: 'fas fa-columns'
+    });
+
+    // general settings
+    let defaultLayout: LayoutType = LayoutType.None;
+    await joplin.settings.registerSetting('defaultLayout', {
+      value: '0',
+      type: SettingItemType.Int,
+      section: 'persistent.layout.settings',
+      isEnum: true,
+      public: true,
+      label: 'Default editor layout',
+      description: 'Default editor layout which is used for all notes that have no "layout" tags specified. If "None" is selected, the current active is kept.',
+      options: {
+        '0': 'None',
+        '1': 'Editor',
+        '2': 'Split View',
+        '3': 'Viewer',
+        '4': 'Rich Text'
+      },
+    });
+
+    async function readSettings(event?: ChangeEvent) {
+      if (event && event.keys.includes('defaultLayout')) {
+        defaultLayout = await SETTINGS.value('defaultLayout');
+      }
+    }
+
+    SETTINGS.onChange(async (event: ChangeEvent) => {
+      await readSettings(event);
+    });
+
+    //#endregion
+
     //#region HELPERS
-
-    enum LayoutType {
-      Editor = 0,
-      Split = 1,
-      Viewer = 2
-    }
-
-    /**
-     * Definition of the favorite descriptions.
-     */
-    interface ILayoutSpec {
-      label: string,
-      panes: string[]
-    }
-
-    /**
-     * Array of favorite descriptions. Order must match with LayoutType enum.
-     */
-    const LayoutSpec: ILayoutSpec[] = [
-      { label: 'layout:editor', panes: ["editor"] }, // Editor
-      { label: 'layout:split', panes: ["editor", "viewer"] }, // Split
-      { label: 'layout:viewer', panes: ["viewer"] } // Viewer
-    ];
 
     async function getAll(path: Path, query: any): Promise<any[]> {
       query.page = 1;
-      let response = await DATA.get(path, query);
+      let response = await joplin.data.get(path, query);
       let result = !!response.items ? response.items : [];
       while (!!response.has_more) {
         query.page += 1;
-        let response = await DATA.get(path, query);
+        let response = await joplin.data.get(path, query);
         result.concat(response.items)
       }
       return result;
@@ -47,20 +63,20 @@ joplin.plugins.register({
 
     async function addTag(noteId: string, layout: LayoutType) {
       let layoutTag = (await getAll(['tags'], { fields: ['id', 'title'], page: 1 }))
-        .find(x => x.title == LayoutSpec[layout].label);
+        .find(x => x.title == layoutDesc[layout].label);
 
       if (!layoutTag) {
-        layoutTag = await DATA.post(['tags'], null, { title: LayoutSpec[layout].label });
+        layoutTag = await joplin.data.post(['tags'], null, { title: layoutDesc[layout].label });
       }
-      await DATA.post(['tags', layoutTag.id, 'notes'], null, { id: noteId });
+      await joplin.data.post(['tags', layoutTag.id, 'notes'], null, { id: noteId });
     }
 
     async function removeTag(noteId: string, layout: LayoutType) {
       let layoutTag = (await getAll(['tags'], { fields: ['id', 'title'], page: 1 }))
-        .find(x => x.title == LayoutSpec[layout].label);
+        .find(x => x.title == layoutDesc[layout].label);
 
       if (layoutTag) {
-        await DATA.delete(["tags", layoutTag.id, "notes", noteId]);
+        await joplin.data.delete(["tags", layoutTag.id, "notes", noteId]);
       }
     }
 
@@ -75,22 +91,36 @@ joplin.plugins.register({
       if (LayoutType.Viewer != layout) {
         await removeTag(noteId, LayoutType.Viewer);
       }
+      if (LayoutType.Richtext != layout) {
+        await removeTag(noteId, LayoutType.Richtext);
+      }
       // add layout tag
       await addTag(noteId, layout);
     }
 
-    async function layoutMatchesVisiblePanes(layout: LayoutType): Promise<boolean> {
+    function visiblePanesMatchLayout(noteVisiblePanes: any[], layout: LayoutType): boolean {
       // noteVisiblePanes = ["editor","viewer"]
-      const noteVisiblePanes: any[] = await SETTINGS.globalValue('noteVisiblePanes');
-      return (LayoutSpec[layout].panes.sort().toString() == noteVisiblePanes.sort().toString());
+      return (layoutDesc[layout].panes.sort().toString() == noteVisiblePanes.sort().toString());
     }
 
     async function toggleVisiblePanes(layout: LayoutType) {
-      for (let i: number = 0; i < 3; i++) {
-        if (await layoutMatchesVisiblePanes(layout)) {
-          break;
+      // console.log(`Toggle layout: ${JSON.stringify(LayoutDesc[layout])}`);
+      const codeView: boolean = await joplin.settings.globalValue('editor.codeView');
+
+      // toggle markdown/rich text editor
+      if (layoutDesc[layout].codeView != codeView) {
+        await joplin.commands.execute('toggleEditors');
+      }
+
+      // toggle panes for markdown editor
+      if (layoutDesc[layout].codeView) {
+        for (let i: number = 0; i < 3; i++) {
+          const visiblePanes: any[] = await joplin.settings.globalValue('noteVisiblePanes');
+          if (visiblePanesMatchLayout(visiblePanes, layout)) {
+            break;
+          }
+          await joplin.commands.execute('toggleVisiblePanes');
         }
-        await COMMANDS.execute('toggleVisiblePanes');
       }
     }
 
@@ -111,18 +141,24 @@ joplin.plugins.register({
         if (!selectedNoteIds) selectedNoteIds = await WORKSPACE.selectedNoteIds();
         if (!selectedNoteIds) return;
 
-        const editor: boolean = await layoutMatchesVisiblePanes(LayoutType.Editor);
-        const split: boolean = await layoutMatchesVisiblePanes(LayoutType.Split);
-        const viewer: boolean = await layoutMatchesVisiblePanes(LayoutType.Viewer);
+        const codeView: boolean = await SETTINGS.globalValue('editor.codeView');
+        const noteVisiblePanes: any[] = await SETTINGS.globalValue('noteVisiblePanes');
+        const editor: boolean = visiblePanesMatchLayout(noteVisiblePanes, LayoutType.Editor);
+        const split: boolean = visiblePanesMatchLayout(noteVisiblePanes, LayoutType.Split);
+        const viewer: boolean = visiblePanesMatchLayout(noteVisiblePanes, LayoutType.Viewer);
 
         // persist for all selected notes
         for (const noteId of selectedNoteIds) {
-          if (editor) {
-            await persistEditorLayout(noteId, LayoutType.Editor);
-          } else if (split) {
-            await persistEditorLayout(noteId, LayoutType.Split);
-          } else if (viewer) {
-            await persistEditorLayout(noteId, LayoutType.Viewer);
+          if (codeView) {
+            if (editor) {
+              await persistEditorLayout(noteId, LayoutType.Editor);
+            } else if (split) {
+              await persistEditorLayout(noteId, LayoutType.Split);
+            } else if (viewer) {
+              await persistEditorLayout(noteId, LayoutType.Viewer);
+            }
+          } else {
+            await persistEditorLayout(noteId, LayoutType.Richtext);
           }
         }
       }
@@ -141,18 +177,23 @@ joplin.plugins.register({
     WORKSPACE.onNoteSelectionChange(async () => {
       try {
         const selectedNote: any = await WORKSPACE.selectedNote();
-        const isCodeEditor: boolean = await SETTINGS.globalValue('editor.codeView');
 
-        if (selectedNote && isCodeEditor) {
+        if (selectedNote) {
           const noteTags = await getAll(['notes', selectedNote.id, 'tags'], { fields: ['id', 'title'], page: 1 });
-          // const tags: any = await DATA.get(['notes', selectedNote.id, 'tags'], { fields: ['title'] });
+          let layout: LayoutType = defaultLayout;
 
-          if (noteTags.find(x => x.title === LayoutSpec[LayoutType.Editor].label)) {        // editor
-            await toggleVisiblePanes(LayoutType.Editor);
-          } else if (noteTags.find(x => x.title === LayoutSpec[LayoutType.Split].label)) {  // split view
-            await toggleVisiblePanes(LayoutType.Split);
-          } else if (noteTags.find(x => x.title === LayoutSpec[LayoutType.Viewer].label)) { // viewer
-            await toggleVisiblePanes(LayoutType.Viewer);
+          if (noteTags.find(x => x.title === layoutDesc[LayoutType.Editor].label)) {          // editor
+            layout = LayoutType.Editor;
+          } else if (noteTags.find(x => x.title === layoutDesc[LayoutType.Split].label)) {    // split view
+            layout = LayoutType.Split;
+          } else if (noteTags.find(x => x.title === layoutDesc[LayoutType.Viewer].label)) {   // viewer
+            layout = LayoutType.Viewer;
+          } else if (noteTags.find(x => x.title === layoutDesc[LayoutType.Richtext].label)) { // rich text
+            layout = LayoutType.Richtext;
+          }
+
+          if (layout > 0) {
+            await toggleVisiblePanes(layout);
           }
         }
       } catch (error) {
@@ -162,5 +203,6 @@ joplin.plugins.register({
 
     //#endregion
 
+    await readSettings();
   }
 });