+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - None
 
+## [1.1.0] - 2021-02-01
+
+### Added
+
+- User option to specify default editor layout ([#3](https://github.com/benji300/joplin-persistent-layout/issues/3))
+- Support to persist rich text editor layout ([#4](https://github.com/benji300/joplin-persistent-layout/issues/4))
+
 ## [1.0.0] - 2021-01-29
 
 - Initial Release